pagebrief-backend-flask/requirements.txt
Flask>=3.0,<4
flask-cors>=4.0,<6
httpx>=0.27,<1
pypdf>=4.2,<7

python-dotenv<2,>=1.0
